refactor header & footer
